Boost integration:
In a kind of a sideways approach, instead of integrating boost with the whole codebase, the only code that will use boost is located within the src/ui directory.
This way, any developer who doesn't wish to deal with boost can simply ignore this location, excluding it from the project.

The general recommendation if you use xcode is to exclude src/ui to keep out of trouble. Basic functionality will/can still be maintained in src/revlanguage in the extent that is needed for debugging etc..

To build all command line tools, including the boost dependent ones, just go to projects/cmake directory and execute these commands:
./regenerate.sh
cmake .
make

The boost libraries are now automatically compiled and linked, and the executables are buildt in the same directory.