Changes to reading of standard data type from NCL
The StandardState labels member will probably lack
the '-' and '?' characters. So setting the states
to these codes should probably be done with special
setGapState and setMissing methods.
The StandardState's stateIndex is problematic because
not all states have an index. in particular the gap
state (treated here as as all 0 bits as other datatypes
treat this "state") and the missing state cannot be
mapped to one index. So this data member has been
removed. The previous accessor now calculates the index
on the fly - and fails with an assertion error if the
state is 0 or has more than one bit "on".