All but EBD tests working
I believe that vectors are the last remaining broken item. Touching a
single element in a vector (where element i is parent to i+1) appears to
not always place all of element i's children in the call to touch(), but
the pre-pass thinks it should. This results in not all calls being
propagated when they should be