Working on BurstEventProposal.
Adding conditions under which the move fails (returns without updating
parameters). It now checks whether the parent or oldest child of the
target node are burst speciations. If so, the move is aborted. This
should prevent polytomies from being formed by nested nodes are assigned
to the burst event.

Also, setting the 'failed' attribute to true in these cases, which
prevents anything from being done when 'undoProposal' is called. This
could be important, because the current state of the node is stored
after these conditions are checked (i.e., the node was not being stored
in these cases, but the state was being "restored" from a previously
stored state if 'undoProposal' is called.